Fix file paths on refactored api folder structure
