+++ server/api/item/Item.model.js
@@ -1,7 +1,7 @@
-var db = require('../config/neo4j').db;
-var spatial = require('../config/neo4j').spatial;
-var spatialURI = require('../config/neo4j').spatialURI;
-var serif = require('../config/neo4j').serif;
+var db = require('../../config/neo4j').db;
+var spatial = require('../../config/neo4j').spatial;
+var spatialURI = require('../../config/neo4j').spatialURI;
+var serif = require('../../config/neo4j').serif;
 var _ = require('lodash');
 
 // serif.query('START n = node(24233) RETURN n', function(err, result){
