+++ server/api/list/List.model.js
@@ -1,4 +1,4 @@
-var db = require('../config/neo4j').db;
+var db = require('../../config/neo4j').db;
 var _ = require('lodash');
 
 var List = function() {};
